[postgres] Fold `values.go` into `adapter.go`
